社会兼职情况、科研情况、培训情况增加pdf上传预览功能

+++ src/main/java/persistence/cadre/CadreBookMapper.java
@@ -2,13 +2,12 @@ package persistence.cadre;
 
 import domain.cadre.CadreBook;
 import domain.cadre.CadreBookExample;
+import java.util.List;
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.session.RowBounds;
 
-import java.util.List;
-
 public interface CadreBookMapper {
-    int countByExample(CadreBookExample example);
+    long countByExample(CadreBookExample example);
 
     int deleteByExample(CadreBookExample example);
