+++ src/main/java/persistence/cadre/CadreResearchMapper.java
@@ -2,13 +2,12 @@ package persistence.cadre;
 
 import domain.cadre.CadreResearch;
 import domain.cadre.CadreResearchExample;
+import java.util.List;
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.session.RowBounds;
 
-import java.util.List;
-
 public interface CadreResearchMapper {
-    int countByExample(CadreResearchExample example);
+    long countByExample(CadreResearchExample example);
 
     int deleteByExample(CadreResearchExample example);
