+++ src/main/java/service/cadre/CadreBookService.java
@@ -52,7 +52,7 @@ public class CadreBookService extends BaseMapper {
             // 干部信息本人直接修改数据校验
             CadreBookExample example = new CadreBookExample();
             example.createCriteria().andCadreIdEqualTo(cadreId).andIdIn(Arrays.asList(ids));
-            int count = cadreBookMapper.countByExample(example);
+            long count = cadreBookMapper.countByExample(example);
             if (count != ids.length) {
                 throw new OpException("参数有误");
             }
